data+model 47f3185b0314 | PR-AUC 0.8945
Closes the traceability gap flagged earlier: nothing linked an MLflow
run to the exact state of gold_machine_hourly_feature that produced
it, which is exactly how the b11 model ended up undocumentable (67 vs
69 vs 78 features depending on when you look at the Postgres schema).

New:
- indusense/data.py: export_gold_dataset() snapshots the Gold table to
  a CSV, hash_file() gives its MD5 — same hash DVC itself computes for
  the .dvc pointer, verified equal (47f3185b0314adc00960476486f177fa
  on both sides).
- indusense/modeling/tracking.py: log_training_run() logs params +
  metrics to MLflow and, when a gold_md5 is given, tags the run with
  it (and the CSV's path) so a run can always be traced back to a
  frozen data snapshot. write_metrics_and_params() writes metrics.json
  / params.yaml (git-tracked, small, diffable) alongside.
- cli.py: `indusense export-gold` (produces the CSV + prints its hash
  and the dvc add command to run) ; `indusense train` gained
  --gold-csv (tags the MLflow run with that file's hash) and
  --metrics-out (writes metrics.json/params.yaml).

Ran the real workflow end to end, not just unit tests:
  uv run indusense export-gold
  dvc add data/gold/gold_dataset.csv && dvc push
  uv run indusense train --gold-csv data/gold/gold_dataset.csv \
      --metrics-out artifacts/models -o artifacts/models/model.joblib
  dvc add artifacts/models/model.joblib && dvc push
This surfaced and fixed a real bug: compute_scale_pos_weight() returned
a numpy.float64, which yaml.safe_dump can't serialize — cast to float()
at the source (isolated fix; generate_model_card.py and the notebooks
compute this inline and were never wired to this helper, so unaffected).
Also caught `dvc push` silently no-op'ing on a brand-new .dvc file not
yet added to git — pushing the target explicitly worked; verified via
the remote folder's contents, not just the CLI's own "up to date" claim.

Model artifact (artifacts/models/model.joblib) is now DVC-tracked, same
pattern as the data. This is new to this codebase — mlruns/b11 remains
Git-tracked as before (untouched) — so it doesn't violate the "one file,
one owner" rule discussed earlier: that rule was about not duplicating
an artifact MLflow already owns, and this run's model was never logged
to MLflow's own artifact store (log_training_run only logs params/
metrics/tags, no mlflow.sklearn.log_model), so DVC is its only owner.

Not done: a `git tag` freeze workflow (documented next, not yet run).

+++ ML/src/indusense/modeling/train.py
@@ -36,7 +36,7 @@ B11_PARAMS = {
 
 
 def compute_scale_pos_weight(y) -> float:
-    return round((y == 0).sum() / (y == 1).sum(), 2)
+    return round(float((y == 0).sum() / (y == 1).sum()), 2)
 
 
 def train_and_evaluate(X_tv, y_tv, X_test, y_test, params: dict) -> tuple[Pipeline, dict]:
